Add support for dispatch on number of output arguments
Now we can dispatch on the number of output args (nargout), by a soft
change in API: the wrapper function must comply with varargout syntax,
instead of defining a single out variable. The old syntax would still
work if none of the concrete implementations has multiple outputs, so
the change should not be breaking backwards compatibility.

Please note though that the way dispatch on nargout is achieved is via
a try-catch statement, which implies more than one concrete call is
attempted while spanning the methodTable.
This could introduce substantial overheads, so maybe we should still
advice to not define multiple outputs or, better, find a cleaner way
to dispatch on them. Anyway all this has to be discussed after some
profiling, which requires to write a thorough test suite.